% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_values.R
\name{depth_values}
\alias{depth_values}
\title{Depth functions wrapper}
\usage{
depth_values(data, u = NULL, method = "Projection", warnings = FALSE)
}
\arguments{
\item{data}{Matrix or Data-Frame of numerical values containing the
observations (rows correspond to observations, columns to variables)}

\item{u}{Matrix or Data-Frame of numerical values containing the coordinates
for which depth values are to be computed (rows correspond to observations,
columns to variables). When missing or NULL, depth values will be computed
for all observations from \code{data}}

\item{method}{Depth function used. Valid options are "L2", "Lui",
"Mahalanobis", "Oja", "Projection" (default), "Spatial", "Tukey"}

\item{warnings}{Logical value, to display the warnings and errors raised by
the underlying depth functions}
}
\value{
Array of numerical values containing the depth of each observation
from \code{data}, or from \code{u} if provided (these values are all set to 0 in the
occurrence of errors)
}
\description{
Computes the depth values with respect to the distribution provided in \code{data}
of either all the coordinates given in \code{u}, or all observations from \code{data}
if \code{u} is not provided. Depth computations rely on the packages
\href{https://CRAN.R-project.org/package=depth}{\code{depth}} and
\href{https://CRAN.R-project.org/package=ddalpha}{\code{ddalpha}}.
}
\examples{

# Illustrative data
data <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2)),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)

# Compute depths
D <- depth_values(data, method = "Projection", warnings = TRUE)

# Plot distribution with depth color scale
plotColors <- colorRampPalette(c("maroon4", "steelblue4", "green4", "gold"))(20)
plot(data, pch = 20, asp = 1, col = plotColors[as.numeric(cut(D, breaks = 20))],
     xlab = "X", ylab = "Y")

# Plot depth values
plot(1:nrow(data), D, pch = 20, col = plotColors[as.numeric(cut(D, breaks = 20))],
     xlab = "Index", ylab = "Depth")

# Compute depth for a single point
depth_values(data, c(10, 3), method = "Projection")

# Compute depth for three sets of coordinates
depth_values(data, rbind(c(10, 3), c(65, 8), c(0, 1)), method = "Projection")

## Data in n>2 dimensions

sigma <- matrix(c(4,2,4,2,4,2,4,2,4), ncol=3)
data <- mvtnorm::rmvnorm(n=500, mean=c(1,2,0), sigma=sigma)
depth_values(data, method = "Projection", warnings = TRUE)

data <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2)),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)
depth_values(data, method = "Projection", warnings = TRUE)

sigma <- matrix(1:25,5)
sigma[lower.tri(sigma)] = t(sigma)[lower.tri(sigma)]
data <- mvtnorm::rmvnorm(n=500, mean=c(1,2,0,0,0), sigma=sigma)
depth_values(data, method = "Projection", warnings = TRUE)

}
\seealso{
\code{\link[=clustBRIC]{clustBRIC()}}, \code{\link[=filter_outliers]{filter_outliers()}}, \code{\link[=median_rec]{median_rec()}}, \code{\link[=median_mv]{median_mv()}}
}
