% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustBRIC.R
\name{plot.BRIC}
\alias{plot.BRIC}
\title{Plot method for \code{BRIC} objects}
\usage{
\method{plot}{BRIC}(
  x,
  contents = "plot",
  showCenters = FALSE,
  col,
  colCenters,
  colClusters = NULL,
  iterationsIndices = NULL,
  iterationsOptions = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{BRIC} (see \code{\link[=clustBRIC]{clustBRIC()}})}

\item{contents}{Contents to be displayed, options are "scatterplot", or
"iterations" (only one option possible)}

\item{showCenters}{Logical value used when \code{contents = "scatterplot"}, to
show or not the clusters' center}

\item{col}{Default color of samples}

\item{colCenters}{Color of the center(s) when \code{contents = "scatterplot"}}

\item{colClusters}{List (or array) of colors for each of the
clusters/iterations (length must be at least equal to the number of groups
identified by the function \code{\link[=clustBRIC]{clustBRIC()}}, i.e. \code{x$nbClusters})}

\item{iterationsIndices}{Numerical value or array of numerical values, used
when \code{contents = "iterations"}, which provides the indices of the
iterations to be plotted. If more than one iteration is requested, an
interactive menu in the console will be used for the selection. 0 or NULL
(default) will include all the iterations. Values that are negative or
superior to the number of iterations performed by the execution
\code{\link[=clustBRIC]{clustBRIC()}} will be ignored}

\item{iterationsOptions}{List of additional parameters to be passed to the
\code{\link[=plot.BRIC.Filtering]{plot.BRIC.Filtering()}} function when \code{contents = "iterations"} is selected
(see \code{\link[=plot.BRIC.Filtering]{plot.BRIC.Filtering()}} for details). Example: \code{iterationsOptions = list(xlab = NA, ylab = NA, contents = c("p.values", "scatterplot"), asp = 1)}}

\item{...}{Other arguments passed to or from other methods (such as pch for
the symbols, main and sub for title and subtitle, xlab, xmin, ...)}
}
\description{
Plot method for \code{BRIC} objects
}
\examples{

# Create a sample distribution and run clustBRIC() function
data <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2)),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)
res <- clustBRIC(data, debug = TRUE)


# Plot the clusters
plot(res)

# Plot the clusters with extra graphic options
plot(res, showCenters = TRUE, main = "Multivariate Mode Estimate",
     col = "gray", colCenters = "black", colClusters = c("yellow","cyan",
     "purple","red"), asp = 1, pch = 3 )

# Plot the second iteration
plot(res, contents = "iteration", iterationsIndices = 2)

# Plot the second iteration (with arguments to plot.BRIC.filtering())
plot(res, contents = "iteration", iterationsIndices = 2,
     iterationsOptions = list(
       contents = c("scatterplot", "p.values"),
       colSelection = "blue", mfrow = c(2,1), asp = 1))

\dontrun{
# Plot all iterations (interactive mode)
plot(res, contents = "iterations")

# Plot the 3 first iterations with options (interactive mode)
plot(res, contents = "iterations", iterationsIndices = c(1:3),
     iterationsOptions = list(
       contents = c("scatterplot"),
       xlim  = c(-50,50),  ylim  = c(-30,30), asp = 1))
}

}
\seealso{
\code{\link[=clustBRIC]{clustBRIC()}}, \code{\link[=print.BRIC]{print.BRIC()}}, \code{\link[=filter_outliers]{filter_outliers()}},
\code{\link[=print.BRIC.Filtering]{print.BRIC.Filtering()}}
}
