% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{print.BRIC.Filtering}
\alias{print.BRIC.Filtering}
\title{Print method for \code{BRIC.Filtering} objects}
\usage{
\method{print}{BRIC.Filtering}(x, maxDisplayed = 200, ...)
}
\arguments{
\item{x}{An object of class \code{BRIC.Filtering} (see \code{\link[=filter_outliers]{filter_outliers()}})}

\item{maxDisplayed}{Number of elements to display in the output (default:
200). Set to NULL (or 0) to show all values.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{BRIC.Filtering} objects
}
\seealso{
\code{\link[=filter_outliers]{filter_outliers()}}, \code{\link[=plot.BRIC.Filtering]{plot.BRIC.Filtering()}}, \code{\link[=clustBRIC]{clustBRIC()}}
}
