% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_rec.R
\name{print.BRIC.MedianRec}
\alias{print.BRIC.MedianRec}
\title{Print method for \code{BRIC.MedianRec} objects}
\usage{
\method{print}{BRIC.MedianRec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BRIC.MedianRec} (see \code{\link[=median_rec]{median_rec()}})}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{BRIC.MedianRec} objects
}
\seealso{
\code{\link[=median_rec]{median_rec()}}, \code{\link[=plot.BRIC.MedianRec]{plot.BRIC.MedianRec()}}
}
