% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustBRIC.R
\name{print.BRIC}
\alias{print.BRIC}
\title{Print method for \code{BRIC} objects}
\usage{
\method{print}{BRIC}(x, maxDisplayed = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{BRIC} (see \code{\link[=clustBRIC]{clustBRIC()}})}

\item{maxDisplayed}{Number of elements to display in the output. Set to NULL
(or 0) to show all values.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{BRIC} objects
}
\seealso{
\code{\link[=clustBRIC]{clustBRIC()}}, \code{\link[=plot.BRIC]{plot.BRIC()}}
}
