% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{test_custom}
\alias{test_custom}
\title{Custom test of a distribution}
\usage{
test_custom(data, test, center = NULL, distances = NULL, warnings = FALSE)
}
\arguments{
\item{data}{Matrix of numerical values containing the observations (one per
row, with columns corresponding to variables)}

\item{test}{Statistical test used (user-defined function)}

\item{center}{Center of the observations from \code{data}}

\item{distances}{Unidimensional array of numerical values (distances of the
observations to \code{center})}

\item{warnings}{Logical value, to display the warnings and errors caught}
}
\value{
p-value of the test
}
\description{
Custom test of a distribution
}
\seealso{
\code{\link[=filter_outliers]{filter_outliers()}}
}
